engine_traits: Move CFHandle methods and types to CFHandleExt trait
